# managify-backend/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Managify project management application. It provides endpoints for managing projects, issues, users, roles, statuses, and project invites.",
        "title": "Managify API",
        "contact": {
            "name": "Doguhan İlter",
            "email": "doguhannilt@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/project-invites": {
            "post": {
                "description": "Creates a new invite for a project.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProjectInvites"
                ],
                "summary": "Create a new project invite",
                "parameters": [
                    {
                        "description": "Project Invite Request",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ProjectInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/project-invites/{id}": {
            "get": {
                "description": "Retrieves project invites by user or project ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProjectInvites"
                ],
                "summary": "Get project invites by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User or Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/project-invites/{inviteId}/respond": {
            "patch": {
                "description": "Accepts or declines a project invite.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProjectInvites"
                ],
                "summary": "Respond to a project invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite ID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action (accept/decline)",
                        "name": "action",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.ProjectInviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
...
    }
}